Validation for endpoints, added comments, clean code

+++ blueprints/restaurant/restaurant_api.py
@@ -1,6 +1,5 @@
 # Restaurant API
-# name and branch must be unique
-from flask import Flask, Blueprint, url_for, request, abort, json, make_response
+from flask import Flask, Blueprint, url_for, request, abort, json
 from bson.objectid import ObjectId
 from pymongo import MongoClient
 
@@ -13,7 +12,6 @@ db = client.entity
 def get_restaurants():
 
 	restaurants = db.entity.find()
-
 	restaurant_list = [i for i in restaurants]
 
 	return json.dumps(restaurant_list)
@@ -21,14 +19,20 @@ def get_restaurants():
 @RESTAURANT_API.route('/add', methods = ['POST'])
 def add_restaurant():
 	if request.form:
+		#required http body
 		name = request.form['name']
 		branch = request.form['branch']
 		operating_hours = request.form['operating_hours']
 		classification = request.form['classification']
 
+		# name and branch must be unique
 		if db.entity.find_one({'name': name, 'branch': branch}):
+
  			return json.dumps({"Duplicate error": "Data cannot be repeated."})
+
  		else:
+
+ 			#set _id to auto incremented numbers
 			db.counters.insert(
 			   {
 			      "_uid": "userid",
@@ -44,6 +48,7 @@ def add_restaurant():
 
 			result = db.entity.insert_one({ "_id": ret, "name": name, "branch": branch, "operating_hours": operating_hours, "classification": classification})
 
+			# insert_one returns acknowledged = true else false
 			if(result.acknowledged):
 				response = json.dumps({"http_status_code": [ 201, "Created" ],"message": "The "+name+" restaurant was successfully added."})
 				return response
@@ -56,19 +61,28 @@ def add_restaurant():
 @RESTAURANT_API.route('/edit', methods = ['POST'])
 def edit_restaurant():
 	if request.form:
+		# required http body
 		name = request.form['name']
 		branch = request.form['branch']
 		operating_hours = request.form['operating_hours']
 		classification = request.form['classification']
 
-		result = db.entity.update_one({"name": name, "branch": branch}, {"$set": { "operating_hours": operating_hours, "classification": classification}})
+		#check if data exist
+		if db.entity.find_one({'name': name, 'branch': branch}):
 
-		if(result.acknowledged):
-			response = json.dumps({"http_status_code": [ 200, "OK" ],"message": "The "+name+" restaurant was successfully updated."})
-			return response
+			# update data from unique name and branch
+			result = db.entity.update_one({"name": name, "branch": branch}, {"$set": { "operating_hours": operating_hours, "classification": classification}})
+
+			# update_one returns acknowledged = true else false
+			if(result.acknowledged):
+				response = json.dumps({"http_status_code": [ 200, "OK" ],"message": "The "+name+" restaurant was successfully updated."})
+				return response
+			else:
+			    response = json.dumps({"message": "Failed to update restaurant."})
+			    return response
+
 		else:
-		    response = json.dumps({"message": "Failed to update restaurant."})
-		    return response
+			return json.dumps({"Error": "Data do not exist."})
 
 	else:
 		return json.dumps({"error": "Please input data."})
@@ -77,17 +91,25 @@ def edit_restaurant():
 def delete_restaurant():
 
 	if request.form:
+		#required http body
 		name = request.form['name']
 		branch = request.form['branch']
 
-		result = db.entity.delete_one({'name' : name, 'branch': branch})
+		#check if data exist
+		if db.entity.find_one({'name': name, 'branch': branch}):
 
-		if(result.acknowledged):
-			response = json.dumps({"http_status_code": [ 200, "OK" ],"message": "The "+name+" restaurant was successfully deleted."})
-			return response
+			# delete data from unique name and branch
+			result = db.entity.delete_one({'name' : name, 'branch': branch})
+
+			# delete_one returns acknowledged = true else false
+			if(result.acknowledged):
+				response = json.dumps({"http_status_code": [ 200, "OK" ],"message": "The "+name+" restaurant was successfully deleted."})
+				return response
+			else:
+			    response = json.dumps({"message": "Failed to delete restaurant."})
+			    return response
 		else:
-		    response = json.dumps({"message": "Failed to delete restaurant."})
-		    return response
+			return json.dumps({"Error": "Data do not exist."})
 
 	else:
 		return json.dumps({"error": "Please input data."})
